Move server code to app/
- Move files (tests and slowmail.go)
- Create `db` command line flag
- Add runtime check that db path is supplied

Until now the db path was hardcoded, and now the server or database can be moved without having to change the hardcoded file path. Now that the code is in `app/`, the go build/go run commands need to be invoked referencing that folder, but nothing else changes. As the codebase grows it can be split across multiple files and later multiple packages.